Improve server configuration options
This implements a new configuration approach, where config settings are
parsed into structs when the service is initialized, and can inform
feature availability throughout the rest of the app.

For example, a BTCPayConfig struct is now parsed on init, and if any
field is missing or invalid, BTCPay would not be available as a billing
option on checkout.

The overall configuration state of the server is displayed in the
console when YeetFile is launched.

+++ web/static/js/account.js
@@ -1,13 +1,30 @@
 document.addEventListener("DOMContentLoaded", () => {
     let logoutBtn = document.getElementById("logout-btn");
     logoutBtn.addEventListener("click", logout);
+
+    let recyclePaymentIDBtn = document.getElementById("recycle-payment-id");
+    recyclePaymentIDBtn.addEventListener("click", recyclePaymentID);
 });
 
 const logout = () => {
-    new YeetFileDB().removeKeys(success => {
-        if (success) {
-            window.location = "/logout";
-        }
-    });
+    let confirmMsg = "Log out of YeetFile?";
+    if (confirm(confirmMsg)) {
+        new YeetFileDB().removeKeys(success => {
+            if (success) {
+                window.location = "/logout";
+            }
+        });
+    }
+}
 
+const recyclePaymentID = () => {
+    let confirmMsg = "Are you sure you want to recycle your payment ID? " +
+        "This will remove all records of past payments you've made.";
+    if (confirm(confirmMsg)) {
+        fetch("/api/recycle-payment-id").then(() => {
+            window.location = "/account";
+        }).catch(() => {
+            alert("Error recycling payment id");
+        });
+    }
 }
